% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam-package.R
\docType{package}
\name{gam-package}
\alias{gam-package}
\title{Generalized Additive Models}
\description{
This package provides functions for fitting and working with generalized additive models as described in chapter 7 of "Statistical Models in S" (Chambers and Hastie (eds), 1991) and "Generalized Additive Models" (Hastie and Tibshirani, 1990).
}
\author{
Trevor Hastie
}
\keyword{models}
\keyword{package}
\keyword{regression}
