\name{gam-internal}
\title{Service functions and as yet undocumented functions for the gam library}
\alias{.First.lib}
\alias{[.smooth}
\alias{general.wam}
\alias{anova.Gamlist}
\alias{as.anova}
\alias{as.data.frame.lo.smooth}
\alias{assign.list}
\alias{Gamlist}
\alias{gam.match}
\alias{gam.nlchisq}
\alias{gam.sp}
\alias{gplot}
\alias{gplot.default}
\alias{gplot.factor}
\alias{gplot.list}
\alias{gplot.matrix}
\alias{gplot.numeric}
\alias{labels.Gam}
\alias{lo.wam}
\alias{newdata.predict.Gam}
\alias{polylo}
\alias{print.Gam}
\alias{print.Gamex}
\alias{print.summary.Gam}
\alias{s.wam}
\alias{ylim.scale}
\description{Internal gam functions}
\author{Trevor Hastie}
\keyword{internal}
