\name{gam-internal}
\title{Service functions and as yet undocumented functions for the gam library}
\alias{.First.lib}
\alias{[.smooth}
\alias{all.wam}
\alias{anova.gamlist}
\alias{as.anova}
\alias{as.data.frame.lo.smooth}
\alias{assign.list}
\alias{deviance.default}
\alias{deviance.glm}
\alias{deviance.lm}
\alias{gamlist}
\alias{gam.match}
\alias{gam.nlchisq}
\alias{gam.random}
\alias{gam.scope}
\alias{gam.slist}
\alias{gam.sp}
\alias{gam.wlist}
\alias{gplot}
\alias{gplot.default}
\alias{gplot.factor}
\alias{gplot.list}
\alias{gplot.matrix}
\alias{gplot.numeric}
\alias{labels.gam}
\alias{newdata.predict.gam}
\alias{polylo}
\alias{print.gam}
\alias{print.summary.gam}
\alias{random}
\alias{ylim.scale}
\description{Internal gam functions}
\author{Trevor Hastie}
\keyword{internal}
